Refactor library so it resides in the lua/surround folder
